sdk/langchain/pyproject.toml
[project]
name = "uipath_langchain"
version = "0.0.62"
description = "UiPath Langchain"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.9"
dependencies = [
    "uipath-sdk>=0.0.75",
    "langgraph>=0.2.70",
    "langchain-core>=0.3.34",
    "langgraph-checkpoint-sqlite>=2.0.3",
    "langchain-community>=0.3.18",
    "langchain-openai>=0.3.3",
    "langchain>=0.3.4",
    "requests>=2.23.3",
    "types-requests>=2.32.0.20241016",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
maintainers = [
    { name = "Marius Cosareanu", email = "marius.cosareanu@uipath.com" },
    { name = "Cristian Pufu", email = "cristian.pufu@uipath.com" }
]

[project.entry-points."uipath_sdk.middlewares"]
register = "uipath_langchain.middlewares:register_middleware"

[project.urls]
Homepage = "https://uipath.com"
Repository = "https://github.com/UiPath/uipath-python"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["uipath_langchain"]

[tool.hatch.build.targets.sdist]
packages = ["uipath_langchain"]

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "ruff>=0.9.4",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pre-commit>=4.1.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
langchain = [
  "uipath-langchain>=0.0.2"
]

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.ruff.lint.per-file-ignores]
"*" = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
plugins = [
  "pydantic.mypy"
]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-ra -q"


